requirements.txt
pinecone-client 
python-dotenv 
openai
pymupdf
langchain
sentence-transformers

